% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{create}
\alias{create}
\title{Create a flights data-only R package}
\usage{
create(path, year, airports, description = getOption("devtools.desc"))
}
\arguments{
\item{path}{Path to the location of the new package}

\item{year}{Year for flights}

\item{airports}{Airports to use for flights}

\item{description}{Options to add to the DESCRIPTION}
}
\description{
Create a flights data-only R package
}

